+++ src/controllers/notes.ts
@@ -3,22 +3,25 @@ import * as serviceNotes from '../services/notes';
 
 
 export async function getAllNotes(req:Request, res:Response){
-    const userId = Number(req.query.userId);
+    if (!req.user) return res.status(401)
+    const userId = Number(req.user.id);
     const result = await serviceNotes.getAllNotesById(userId)
 
     res.status(200).send(result)
 }
 
 export async function getNote(req:Request, res:Response){
+    if (!req.user) return res.status(401)
     const noteId = Number(req.params.id);
-    const userId = Number(req.query.userId);
+    const userId = Number(req.user.id);
     const result = await serviceNotes.getNoteById(userId, noteId);
 
     res.status(200).send(result)
 }
 
 export async function createNote(req:Request, res:Response){
-    const userId = Number(req.query.userId);
+    if (!req.user) return res.status(401)
+    const userId = Number(req.user.id);
     const {contactId, content } = req.body;
     const result = await serviceNotes.createNote(userId, contactId, content);
 
@@ -26,8 +29,9 @@ export async function createNote(req:Request, res:Response){
 }
 
 export async function updateNote(req:Request, res:Response){
+    if (!req.user) return res.status(401)
     const noteId = Number(req.params.id);
-    const userId = Number(req.query.userId);
+    const userId = Number(req.user.id);
     const content = String(req.body.content);
     const result = await serviceNotes.updateNote(userId, noteId, content);
 
@@ -35,8 +39,9 @@ export async function updateNote(req:Request, res:Response){
 }
 
 export async function deleteNote(req:Request, res:Response){
+    if (!req.user) return res.status(401)
     const noteId = Number(req.params.id);
-    const userId = Number(req.query.userId);
+    const userId = Number(req.user.id);
 
     const result = await serviceNotes.deleteNote(userId, noteId);
     res.status(200).send(result)
